* added CVS tags to all files.  why weren't they there before?
* fixed a bug (identified by Will Crain) in which an IMG tag with a
USEMAP attribute does the wrong thing.  it seems HTML::LinkExtor returns
a five element list in this case, instead of the three elements that
i was expecting.  hence, using -1 as an index gives the fifth element
when i really wanted the third one.  i've tried a couple of test cases
and haven't broken this and it does not appear to depend on position in
the HTML.

there has to be a better way to do this though.


git-svn-id: http://brian-d-foy.svn.sourceforge.net/svnroot/brian-d-foy/trunk@39 d9f60a02-fc28-0410-9752-bc0d6c4f57a6