* rt.cpan.org: Ticket #16942
For HTML tags that could contain multiple attributes that had URLs
as values (e.g. IMG with SRC and USEMAP), only the last URL was
resolved to a full URL if a BASE HREF was found. I've fixed that
so that all attributes which I expect to have a URL value are
resolved.


git-svn-id: http://brian-d-foy.svn.sourceforge.net/svnroot/brian-d-foy/trunk@1726 d9f60a02-fc28-0410-9752-bc0d6c4f57a6